backend/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
pydantic-settings==2.6.1
pydantic==2.10.3
python-multipart==0.0.18
httpx==0.28.1
cryptography==44.0.0
bcrypt==4.2.1
PyJWT==2.10.1
python-dotenv==1.0.1
